+++ includes/libs/filebackend/SwiftFileBackend.php
@@ -22,6 +22,7 @@
  * @author Russ Nelson
  */
 
+use MediaWiki\Utils\MWTimestamp;
 use Psr\Log\LoggerInterface;
 use Wikimedia\AtEase\AtEase;
 use Wikimedia\RequestTimeout\TimeoutException;
@@ -303,13 +304,13 @@ class SwiftFileBackend extends FileBackendStore {
 
 		$method = __METHOD__;
 		$handler = function ( array $request, StatusValue $status ) use ( $method, $params ) {
-			[ $rcode, $rdesc, , , $rerr ] = $request['response'];
+			[ $rcode, $rdesc, , $rbody, $rerr ] = $request['response'];
 			if ( $rcode === 201 || $rcode === 202 ) {
 				// good
 			} elseif ( $rcode === 412 ) {
 				$status->fatal( 'backend-fail-contenttype', $params['dst'] );
 			} else {
-				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc );
+				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc, $rbody );
 			}
 
 			return SwiftFileOpHandle::CONTINUE_IF_OK;
@@ -390,13 +391,13 @@ class SwiftFileBackend extends FileBackendStore {
 
 		$method = __METHOD__;
 		$handler = function ( array $request, StatusValue $status ) use ( $method, $params ) {
-			[ $rcode, $rdesc, , , $rerr ] = $request['response'];
+			[ $rcode, $rdesc, , $rbody, $rerr ] = $request['response'];
 			if ( $rcode === 201 || $rcode === 202 ) {
 				// good
 			} elseif ( $rcode === 412 ) {
 				$status->fatal( 'backend-fail-contenttype', $params['dst'] );
 			} else {
-				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc );
+				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc, $rbody );
 			}
 
 			return SwiftFileOpHandle::CONTINUE_IF_OK;
@@ -445,7 +446,7 @@ class SwiftFileBackend extends FileBackendStore {
 
 		$method = __METHOD__;
 		$handler = function ( array $request, StatusValue $status ) use ( $method, $params ) {
-			[ $rcode, $rdesc, , , $rerr ] = $request['response'];
+			[ $rcode, $rdesc, , $rbody, $rerr ] = $request['response'];
 			if ( $rcode === 201 ) {
 				// good
 			} elseif ( $rcode === 404 ) {
@@ -453,7 +454,7 @@ class SwiftFileBackend extends FileBackendStore {
 					$status->fatal( 'backend-fail-copy', $params['src'], $params['dst'] );
 				}
 			} else {
-				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc );
+				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc, $rbody );
 			}
 
 			return SwiftFileOpHandle::CONTINUE_IF_OK;
@@ -507,7 +508,7 @@ class SwiftFileBackend extends FileBackendStore {
 
 		$method = __METHOD__;
 		$handler = function ( array $request, StatusValue $status ) use ( $method, $params ) {
-			[ $rcode, $rdesc, , , $rerr ] = $request['response'];
+			[ $rcode, $rdesc, , $rbody, $rerr ] = $request['response'];
 			if ( $request['method'] === 'PUT' && $rcode === 201 ) {
 				// good
 			} elseif ( $request['method'] === 'DELETE' && $rcode === 204 ) {
@@ -520,7 +521,7 @@ class SwiftFileBackend extends FileBackendStore {
 					return SwiftFileOpHandle::CONTINUE_NO;
 				}
 			} else {
-				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc );
+				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc, $rbody );
 			}
 
 			return SwiftFileOpHandle::CONTINUE_IF_OK;
@@ -554,7 +555,7 @@ class SwiftFileBackend extends FileBackendStore {
 
 		$method = __METHOD__;
 		$handler = function ( array $request, StatusValue $status ) use ( $method, $params ) {
-			[ $rcode, $rdesc, , , $rerr ] = $request['response'];
+			[ $rcode, $rdesc, , $rbody, $rerr ] = $request['response'];
 			if ( $rcode === 204 ) {
 				// good
 			} elseif ( $rcode === 404 ) {
@@ -562,7 +563,7 @@ class SwiftFileBackend extends FileBackendStore {
 					$status->fatal( 'backend-fail-delete', $params['src'] );
 				}
 			} else {
-				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc );
+				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc, $rbody );
 			}
 
 			return SwiftFileOpHandle::CONTINUE_IF_OK;
@@ -617,13 +618,13 @@ class SwiftFileBackend extends FileBackendStore {
 
 		$method = __METHOD__;
 		$handler = function ( array $request, StatusValue $status ) use ( $method, $params ) {
-			[ $rcode, $rdesc, , , $rerr ] = $request['response'];
+			[ $rcode, $rdesc, , $rbody, $rerr ] = $request['response'];
 			if ( $rcode === 202 ) {
 				// good
 			} elseif ( $rcode === 404 ) {
 				$status->fatal( 'backend-fail-describe', $params['src'] );
 			} else {
-				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc );
+				$this->onError( $status, $method, $params, $rerr, $rcode, $rdesc, $rbody );
 			}
 		};
 
@@ -650,12 +651,8 @@ class SwiftFileBackend extends FileBackendStore {
 		} elseif ( $stat === self::$RES_ERROR ) {
 			$status->fatal( 'backend-fail-internal', $this->name );
 			$this->logger->error( __METHOD__ . ': cannot get container stat' );
-
-			return $status;
-		}
-
-		// (b) Create container as needed with proper ACLs
-		if ( $stat === false ) {
+		} else {
+			// (b) Create container as needed with proper ACLs
 			$params['op'] = 'prepare';
 			$status->merge( $this->createContainer( $fullCont, $params ) );
 		}
@@ -679,7 +676,7 @@ class SwiftFileBackend extends FileBackendStore {
 				$readUsers,
 				$writeUsers
 			) );
-		} elseif ( $stat === false ) {
+		} elseif ( $stat === self::$RES_ABSENT ) {
 			$status->fatal( 'backend-fail-usable', $params['dir'] );
 		} else {
 			$status->fatal( 'backend-fail-internal', $this->name );
@@ -703,7 +700,7 @@ class SwiftFileBackend extends FileBackendStore {
 				$readUsers,
 				$writeUsers
 			) );
-		} elseif ( $stat === false ) {
+		} elseif ( $stat === self::$RES_ABSENT ) {
 			$status->fatal( 'backend-fail-usable', $params['dir'] );
 		} else {
 			$status->fatal( 'backend-fail-internal', $this->name );
@@ -723,17 +720,13 @@ class SwiftFileBackend extends FileBackendStore {
 
 		// (a) Check the container
 		$stat = $this->getContainerStat( $fullCont, true );
-		if ( $stat === false ) {
+		if ( $stat === self::$RES_ABSENT ) {
 			return $status; // ok, nothing to do
-		} elseif ( !is_array( $stat ) ) {
+		} elseif ( $stat === self::$RES_ERROR ) {
 			$status->fatal( 'backend-fail-internal', $this->name );
 			$this->logger->error( __METHOD__ . ': cannot get container stat' );
-
-			return $status;
-		}
-
-		// (b) Delete the container if empty
-		if ( $stat['count'] == 0 ) {
+		} elseif ( is_array( $stat ) && $stat['count'] == 0 ) {
+			// (b) Delete the container if empty
 			$params['op'] = 'clean';
 			$status->merge( $this->deleteContainer( $fullCont, $params ) );
 		}
@@ -865,7 +858,7 @@ class SwiftFileBackend extends FileBackendStore {
 		] + self::DEFAULT_HTTP_OPTIONS;
 		$reqs = $this->http->runMulti( $reqs, $opts );
 		foreach ( $reqs as $path => $op ) {
-			[ $rcode, $rdesc, $rhdrs, , $rerr ] = $op['response'];
+			[ $rcode, $rdesc, $rhdrs, $rbody, $rerr ] = $op['response'];
 			if ( $rcode >= 200 && $rcode <= 299 ) {
 				rewind( $op['stream'] ); // start from the beginning
 				$content = (string)stream_get_contents( $op['stream'] );
@@ -884,7 +877,7 @@ class SwiftFileBackend extends FileBackendStore {
 			} else {
 				$contents[$path] = self::$RES_ERROR;
 				$this->onError( null, __METHOD__,
-					[ 'src' => $path ] + $ep, $rerr, $rcode, $rdesc );
+					[ 'src' => $path ] + $ep, $rerr, $rcode, $rdesc, $rbody );
 			}
 			fclose( $op['stream'] ); // close open handle
 		}
@@ -929,10 +922,10 @@ class SwiftFileBackend extends FileBackendStore {
 	 *
 	 * @param string $fullCont Resolved container name
 	 * @param string $dir Resolved storage directory with no trailing slash
-	 * @param string|null &$after Resolved container relative path to list items after
+	 * @param string|null &$after Resolved container relative path used for continuation paging
 	 * @param int $limit Max number of items to list
-	 * @param array $params Parameters for getDirectoryList()
-	 * @return array List of container relative resolved paths of directories directly under $dir
+	 * @param array $params Parameters for {@link getDirectoryList()}
+	 * @return string[] List of resolved container relative directories directly under $dir
 	 * @throws FileBackendError
 	 */
 	public function getDirListPageInternal( $fullCont, $dir, &$after, $limit, array $params ) {
@@ -1014,12 +1007,12 @@ class SwiftFileBackend extends FileBackendStore {
 	 * @param string $dir Resolved storage directory with no trailing slash
 	 * @param string|null &$after Resolved container relative path of file to list items after
 	 * @param int $limit Max number of items to list
-	 * @param array $params Parameters for getDirectoryList()
-	 * @return array List of resolved container relative paths of files under $dir
+	 * @param array $params Parameters for {@link getFileList()}
+	 * @return array[] List of (name, stat map or null) tuples under $dir
 	 * @throws FileBackendError
 	 */
 	public function getFileListPageInternal( $fullCont, $dir, &$after, $limit, array $params ) {
-		$files = []; // list of (path, stat array or null) entries
+		$files = []; // list of (path, stat map or null) entries
 		if ( $after === INF ) {
 			return $files; // nothing more
 		}
@@ -1045,7 +1038,7 @@ class SwiftFileBackend extends FileBackendStore {
 			}
 		}
 
-		// Reformat this list into a list of (name, stat array or null) entries
+		// Reformat this list into a list of (name, stat map or null) entries
 		if ( !$status->isOK() ) {
 			throw new FileBackendError( "Iterator page I/O error." );
 		}
@@ -1069,7 +1062,7 @@ class SwiftFileBackend extends FileBackendStore {
 	 * and extracting any stat info if provided in $objects
 	 *
 	 * @param stdClass[]|string[] $objects List of stdClass items or object names
-	 * @return array List of (names,stat array or null) entries
+	 * @return array[] List of (name, stat map or null) entries
 	 */
 	private function buildFileObjectListing( array $objects ) {
 		$names = [];
@@ -1176,7 +1169,7 @@ class SwiftFileBackend extends FileBackendStore {
 		}
 
 		$handle = fopen( 'php://output', 'wb' );
-		[ $rcode, $rdesc, , , $rerr ] = $this->http->run( [
+		[ $rcode, $rdesc, , $rbody, $rerr ] = $this->http->run( [
 			'method' => 'GET',
 			'url' => $this->storageUrl( $auth, $srcCont, $srcRel ),
 			'headers' => $this->authTokenHeaders( $auth )
@@ -1195,7 +1188,7 @@ class SwiftFileBackend extends FileBackendStore {
 			$this->clearCache( [ $params['src'] ] );
 			$this->deleteFileCache( $params['src'] );
 		} else {
-			$this->onError( $status, __METHOD__, $params, $rerr, $rcode, $rdesc );
+			$this->onError( $status, __METHOD__, $params, $rerr, $rcode, $rdesc, $rbody );
 		}
 
 		return $status;
@@ -1241,7 +1234,7 @@ class SwiftFileBackend extends FileBackendStore {
 		] + self::DEFAULT_HTTP_OPTIONS;
 		$reqs = $this->http->runMulti( $reqs, $opts );
 		foreach ( $reqs as $path => $op ) {
-			[ $rcode, $rdesc, $rhdrs, , $rerr ] = $op['response'];
+			[ $rcode, $rdesc, $rhdrs, $rbody, $rerr ] = $op['response'];
 			fclose( $op['stream'] ); // close open handle
 			if ( $rcode >= 200 && $rcode <= 299 ) {
 				/** @var TempFSFile $tmpFile */
@@ -1268,7 +1261,7 @@ class SwiftFileBackend extends FileBackendStore {
 			} else {
 				$tmpFiles[$path] = self::$RES_ERROR;
 				$this->onError( null, __METHOD__,
-					[ 'src' => $path ] + $ep, $rerr, $rcode, $rdesc );
+					[ 'src' => $path ] + $ep, $rerr, $rcode, $rdesc, $rbody );
 			}
 		}
 
@@ -1459,12 +1452,12 @@ class SwiftFileBackend extends FileBackendStore {
 	}
 
 	/**
-	 * Get a Swift container stat array, possibly from process cache.
+	 * Get a Swift container stat map, possibly from process cache.
 	 * Use $reCache if the file count or byte count is needed.
 	 *
 	 * @param string $container Container name
 	 * @param bool $bypassCache Bypass all caches and load from Swift
-	 * @return array|bool|null False on 404, null on failure
+	 * @return array|false|null False on 404, null on failure
 	 */
 	protected function getContainerStat( $container, $bypassCache = false ) {
 		/** @noinspection PhpUnusedLocalVariableInspection */
@@ -1481,7 +1474,7 @@ class SwiftFileBackend extends FileBackendStore {
 				return self::$RES_ERROR;
 			}
 
-			[ $rcode, $rdesc, $rhdrs, , $rerr ] = $this->http->run( [
+			[ $rcode, $rdesc, $rhdrs, $rbody, $rerr ] = $this->http->run( [
 				'method' => 'HEAD',
 				'url' => $this->storageUrl( $auth, $container ),
 				'headers' => $this->authTokenHeaders( $auth )
@@ -1502,7 +1495,7 @@ class SwiftFileBackend extends FileBackendStore {
 				return self::$RES_ABSENT;
 			} else {
 				$this->onError( null, __METHOD__,
-					[ 'cont' => $container ], $rerr, $rcode, $rdesc );
+					[ 'cont' => $container ], $rerr, $rcode, $rdesc, $rbody );
 
 				return self::$RES_ERROR;
 			}
@@ -1539,7 +1532,7 @@ class SwiftFileBackend extends FileBackendStore {
 			$writeUsers = array_merge( $this->secureWriteUsers, [ $this->swiftUser ] );
 		}
 
-		[ $rcode, $rdesc, , , $rerr ] = $this->http->run( [
+		[ $rcode, $rdesc, , $rbody, $rerr ] = $this->http->run( [
 			'method' => 'PUT',
 			'url' => $this->storageUrl( $auth, $container ),
 			'headers' => $this->authTokenHeaders( $auth ) + [
@@ -1553,7 +1546,7 @@ class SwiftFileBackend extends FileBackendStore {
 		} elseif ( $rcode === 202 ) { // already there
 			// this shouldn't really happen, but is OK
 		} else {
-			$this->onError( $status, __METHOD__, $params, $rerr, $rcode, $rdesc );
+			$this->onError( $status, __METHOD__, $params, $rerr, $rcode, $rdesc, $rbody );
 		}
 
 		return $status;
@@ -1576,7 +1569,7 @@ class SwiftFileBackend extends FileBackendStore {
 			return $status;
 		}
 
-		[ $rcode, $rdesc, , , $rerr ] = $this->http->run( [
+		[ $rcode, $rdesc, , $rbody, $rerr ] = $this->http->run( [
 			'method' => 'DELETE',
 			'url' => $this->storageUrl( $auth, $container ),
 			'headers' => $this->authTokenHeaders( $auth )
@@ -1589,8 +1582,7 @@ class SwiftFileBackend extends FileBackendStore {
 		} elseif ( $rcode === 409 ) { // not empty
 			$this->onError( $status, __METHOD__, $params, $rerr, $rcode, $rdesc ); // race?
 		} else {
-			// @phan-suppress-previous-line PhanPluginDuplicateIfStatements
-			$this->onError( $status, __METHOD__, $params, $rerr, $rcode, $rdesc );
+			$this->onError( $status, __METHOD__, $params, $rerr, $rcode, $rdesc, $rbody );
 		}
 
 		return $status;
@@ -1653,7 +1645,7 @@ class SwiftFileBackend extends FileBackendStore {
 		} elseif ( $rcode === 404 ) {
 			$status->value = []; // no container
 		} else {
-			$this->onError( $status, __METHOD__, $params, $rerr, $rcode, $rdesc );
+			$this->onError( $status, __METHOD__, $params, $rerr, $rcode, $rdesc, $rbody );
 		}
 
 		return $status;
@@ -1683,7 +1675,7 @@ class SwiftFileBackend extends FileBackendStore {
 			if ( $cstat === self::$RES_ABSENT ) {
 				$stats[$path] = self::$RES_ABSENT;
 				continue; // ok, nothing to do
-			} elseif ( !is_array( $cstat ) ) {
+			} elseif ( $cstat === self::$RES_ERROR ) {
 				$stats[$path] = self::$RES_ERROR;
 				continue;
 			}
@@ -1701,13 +1693,13 @@ class SwiftFileBackend extends FileBackendStore {
 		] + self::DEFAULT_HTTP_OPTIONS;
 		$reqs = $this->http->runMulti( $reqs, $opts );
 		foreach ( $reqs as $path => $op ) {
-			[ $rcode, $rdesc, $rhdrs, , $rerr ] = $op['response'];
+			[ $rcode, $rdesc, $rhdrs, $rbody, $rerr ] = $op['response'];
 			if ( $rcode === 200 || $rcode === 204 ) {
 				// Update the object if it is missing some headers
 				if ( !empty( $params['requireSHA1'] ) ) {
 					$rhdrs = $this->addMissingHashMetadata( $rhdrs, $path );
 				}
-				// Load the stat array from the headers
+				// Load the stat map from the headers
 				$stat = $this->getStatFromHeaders( $rhdrs );
 				if ( $this->isRGW ) {
 					$stat['latest'] = true; // strong consistency
@@ -1716,7 +1708,7 @@ class SwiftFileBackend extends FileBackendStore {
 				$stat = self::$RES_ABSENT;
 			} else {
 				$stat = self::$RES_ERROR;
-				$this->onError( null, __METHOD__, $params, $rerr, $rcode, $rdesc );
+				$this->onError( null, __METHOD__, $params, $rerr, $rcode, $rdesc, $rbody );
 			}
 			$stats[$path] = $stat;
 		}
@@ -1770,7 +1762,7 @@ class SwiftFileBackend extends FileBackendStore {
 				// Skew the timestamp for worst case to avoid using stale credentials
 				$this->authSessionTimestamp = time() - (int)ceil( $this->authTTL / 2 );
 			} else { // cache miss
-				[ $rcode, , $rhdrs, , ] = $this->http->run( [
+				[ $rcode, , $rhdrs, $rbody, ] = $this->http->run( [
 					'method' => 'GET',
 					'url' => "{$this->swiftAuthUrl}/v1.0",
 					'headers' => [
@@ -1793,14 +1785,14 @@ class SwiftFileBackend extends FileBackendStore {
 
 					return null;
 				} else {
-					$this->onError( null, __METHOD__, [], "HTTP return code: $rcode", $rcode );
+					$this->onError( null, __METHOD__, [], "HTTP return code: $rcode", $rcode, $rbody );
 					$this->authErrorTimestamp = time();
 
 					return null;
 				}
 			}
 			// Ceph RGW does not use <account> in URLs (OpenStack Swift uses "/v1/<account>")
-			if ( substr( $this->authCreds['storage_url'], -3 ) === '/v1' ) {
+			if ( str_ends_with( $this->authCreds['storage_url'], '/v1' ) ) {
 				$this->isRGW = true; // take advantage of strong consistency in Ceph
 			}
 		}
@@ -1816,10 +1808,10 @@ class SwiftFileBackend extends FileBackendStore {
 	 */
 	protected function storageUrl( array $creds, $container = null, $object = null ) {
 		$parts = [ $creds['storage_url'] ];
-		if ( strlen( $container ) ) {
+		if ( strlen( $container ?? '' ) ) {
 			$parts[] = rawurlencode( $container );
 		}
-		if ( strlen( $object ) ) {
+		if ( strlen( $object ?? '' ) ) {
 			$parts[] = str_replace( "%2F", "/", rawurlencode( $object ) );
 		}
 
@@ -1854,8 +1846,9 @@ class SwiftFileBackend extends FileBackendStore {
 	 * @param string $err Error string
 	 * @param int $code HTTP status
 	 * @param string $desc HTTP StatusValue description
+	 * @param string $body HTTP body
 	 */
-	public function onError( $status, $func, array $params, $err = '', $code = 0, $desc = '' ) {
+	public function onError( $status, $func, array $params, $err = '', $code = 0, $desc = '', $body = '' ) {
 		if ( $status instanceof StatusValue ) {
 			$status->fatal( 'backend-fail-internal', $this->name );
 		}
@@ -1872,6 +1865,10 @@ class SwiftFileBackend extends FileBackendStore {
 		if ( $err ) {
 			$msg .= ': {err}';
 			$msgParams['err'] = $err;
+		}
+		if ( $code == 502 ) {
+			$msg .= ' ({truncatedBody})';
+			$msgParams['truncatedBody'] = substr( strip_tags( $body ), 0, 100 );
 		}
 		$this->logger->error( $msg, $msgParams );
 	}
